Move PaginationInfo and PaginationResult into the pagination package.
