+++ server/app/controllers/admin/AdminApplicationController.java
@@ -34,7 +34,6 @@ import repository.SubmittedApplicationFilter;
 import repository.TimeFilter;
 import repository.VersionRepository;
 import services.DateConverter;
-import services.PaginationResult;
 import services.UrlUtils;
 import services.applicant.AnswerData;
 import services.applicant.ApplicantService;
@@ -49,6 +48,7 @@ import services.export.CsvExporterService;
 import services.export.JsonExporterService;
 import services.export.PdfExporter;
 import services.pagination.PageNumberPaginationSpec;
+import services.pagination.PaginationResult;
 import services.pagination.SubmitTimeSequentialAccessPaginationSpec;
 import services.program.ProgramDefinition;
 import services.program.ProgramNotFoundException;
