+++ server/app/controllers/api/ProgramApplicationsApiController.java
@@ -22,8 +22,8 @@ import repository.SubmittedApplicationFilter;
 import repository.TimeFilter;
 import repository.VersionRepository;
 import services.DateConverter;
-import services.PaginationResult;
 import services.export.JsonExporterService;
+import services.pagination.PaginationResult;
 import services.pagination.RowIdSequentialAccessPaginationSpec;
 import services.program.ProgramNotFoundException;
 import services.program.ProgramService;
