+++ server/app/controllers/ti/TrustedIntermediaryController.java
@@ -23,10 +23,10 @@ import play.mvc.Http;
 import play.mvc.Result;
 import repository.AccountRepository;
 import repository.SearchParameters;
-import services.PaginationInfo;
 import services.applicant.ApplicantPersonalInfo;
 import services.applicant.ApplicantPersonalInfo.Representation;
 import services.applicant.exception.ApplicantNotFoundException;
+import services.pagination.PaginationInfo;
 import services.ti.AddNewApplicantReturnObject;
 import services.ti.TrustedIntermediarySearchResult;
 import services.ti.TrustedIntermediaryService;
