+++ server/app/repository/ProgramRepository.java
@@ -30,10 +30,10 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import play.cache.NamedCache;
 import play.cache.SyncCacheApi;
-import services.PaginationResult;
 import services.Path;
 import services.WellKnownPaths;
 import services.pagination.BasePaginationSpec;
+import services.pagination.PaginationResult;
 import services.program.BlockDefinition;
 import services.program.ProgramDefinition;
 import services.program.ProgramDraftNotFoundException;
