+++ server/app/services/export/JsonExporterService.java
@@ -19,7 +19,6 @@ import org.apache.commons.lang3.NotImplementedException;
 import repository.SubmittedApplicationFilter;
 import services.CfJsonDocumentContext;
 import services.DateConverter;
-import services.PaginationResult;
 import services.Path;
 import services.applicant.ApplicantData;
 import services.applicant.ApplicantService;
@@ -27,6 +26,7 @@ import services.applicant.JsonPathProvider;
 import services.applicant.question.ApplicantQuestion;
 import services.export.enums.RevisionState;
 import services.export.enums.SubmitterType;
+import services.pagination.PaginationResult;
 import services.pagination.SubmitTimeSequentialAccessPaginationSpec;
 import services.program.ProgramDefinition;
 import services.program.ProgramService;
