+++ server/app/services/pagination/PaginationResult.java
@@ -1,10 +1,8 @@
-package services;
+package services.pagination;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
 import com.google.common.collect.ImmutableList;
-
-// TODO(#8841): Move this class into the pagination subpackage.
 
 /** Contains a single page of results and pagination state for a paginated query. */
 public class PaginationResult<T> {
